Rename to _DomainsContainer
